refactor(perl-dead-code): split dead-branch and report into SRP modules (#0000)
